BLOC_04/AIRFLOW/requirements.txt

# Provider SQL commun (obligatoire pour SQLExecuteQueryOperator)
apache-airflow-providers-common-sql

# Provider PostgreSQL
apache-airflow-providers-postgres

# Autres librairies utiles
pandas>=2.0.0
requests>=2.32.0
scikit-learn>=1.3.0
python-dotenv
boto3
mlflow==2.21.3
xgboost>=2.0.0
prophet

# SMTP

apache-airflow-providers-smtp

# Evidently
evidently>=0.4.0,<0.5.0

matplotlib==3.9.2
seaborn==0.13.2